+++ src/adservice/.gitignore
@@ -6,3 +6,4 @@
 build/**
 
 
+/.gradle/
