+++ src/ratingService/src/main/java/hipstershop/RatingService.java
@@ -1,20 +1,22 @@
-package cse.project.grpcserver;
+package hipstershop;
 
 import java.io.IOException;
 
-import cse.project.ratingservice.RatingService;
+import cse.project.ratingservice.RatingServiceRate;
+import io.opencensus.contrib.grpc.metrics.RpcViews;
 import io.grpc.health.v1.HealthCheckResponse.ServingStatus;
 import io.grpc.services.*;
 import io.grpc.Server;
 import io.grpc.ServerBuilder;
 
-public class GrpcServer {
+public class RatingService {
 	public static void main(String[] args) throws IOException, InterruptedException {
 
+		RpcViews.registerAllViews();
 		Server server;
 		HealthStatusManager healthMgr;
 		healthMgr = new HealthStatusManager();
-		server = ServerBuilder.forPort(9090).addService(new RatingService()).addService(healthMgr.getHealthService()).build();
+		server = ServerBuilder.forPort(9090).addService(new RatingServiceRate()).addService(healthMgr.getHealthService()).build();
 		server.start();
 		System.out.print("Service started at "+server.getPort());
 		healthMgr.setStatus("", ServingStatus.SERVING);
